requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
aiofiles>=23.2.0
schedule>=1.2.0
openai>=1.3.0
selenium>=4.15.0
playwright>=1.40.0
cryptography>=41.0.0
jinja2>=3.1.0
